requirements.txt
cookidoo-api>=0.15.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
